Refactor TV card UI to modular CardGrid components
- Changed card size as it was too large and only a few channels were visible on screen
- Added a demo EPG to the main card UI on focus (actual EPG data will be added later)
- Replaced `CardChannelLayout` and `CardChannelLayoutM3U` with new modular `CardGrid` and `CardGridM3U` components
- Improve code organization and maintainability for TV card UIs
